requirements/requirements-test.txt
pytest==3.9.1
pytest-django==3.4.3
factory-boy
